Example updated for RN 0.40.0
