--- Example/android/.gitignore
@@ -1,55 +1,0 @@
-# Built application files
-*.apk
-*.ap_
-
-# Files for the Dalvik VM
-*.dex
-
-# Java class files
-*.class
-
-# Generated files
-bin/
-gen/
-out/
-
-# Gradle files
-.gradle/
-build/
-
-# Local configuration file (sdk path, etc)
-local.properties
-
-# Proguard folder generated by Eclipse
-proguard/
-
-# Log Files
-*.log
-
-# Android Studio Navigation editor temp files
-.navigation/
-
-# Android Studio captures folder
-captures/
-
-# Intellij
-*.iml
-
-# Keystore files
-*.jks
-
-# Eclipse Metadata
-.metadata/
-#
-# Mac OS X clutter
-*.DS_Store
-#
-# Windows clutter
-Thumbs.db
-#
-# # Intellij IDEA (see https://intellij-support.jetbrains.com/entries/23393067)
-*.iws
-.idea/libraries
-.idea/tasks.xml
-.idea/vcs.xml
-.idea/workspace.xml
