+++ Example/android/settings.gradle
@@ -1,5 +1,5 @@
 rootProject.name = 'Example'
-
-include ':app'
 include ':react-native-spinkit'
 project(':react-native-spinkit').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-spinkit/android')
+
+include ':app'
