+++ Example/index.ios.js
@@ -1,7 +1,6 @@
 var React = require('react');
 var ReactNative = require('react-native');
 
-
 var {
   AppRegistry,
   StyleSheet,
